+++ src/gui_app/folder_browser/drag_drop.rs
@@ -1,13 +1,9 @@
 use radiant::{gui::types::Point, prelude as ui, widgets::DragHandleMessage};
-use std::{
-    collections::HashSet,
-    fs,
-    path::{Path, PathBuf},
-};
+use std::path::{Path, PathBuf};
 
 use super::{
     FolderBrowserDrag, FolderBrowserState, FolderDragPreview, FolderDropResult,
-    path_helpers::file_label, plural,
+    path_helpers::file_label,
 };
 
 impl FolderBrowserState {
@@ -151,145 +147,5 @@ impl FolderBrowserState {
             }),
             None => false,
         }
-    }
-
-    fn move_folder_to_folder(
-        &mut self,
-        folder_id: &str,
-        target_folder_id: &str,
-    ) -> Result<FolderDropResult, String> {
-        if self.rename_active() {
-            return Err(String::from("Finish rename before moving a folder"));
-        }
-        if self.selected_folder_is_source_root_id(folder_id) {
-            return Err(String::from("Root folder cannot be moved"));
-        }
-        let source_folder = self
-            .find_folder(folder_id)
-            .cloned()
-            .ok_or_else(|| String::from("Folder move failed: source folder is missing"))?;
-        let target_folder = self
-            .find_folder(target_folder_id)
-            .cloned()
-            .ok_or_else(|| String::from("Folder move failed: target folder is missing"))?;
-        let old_path = PathBuf::from(&source_folder.id);
-        let target_path = PathBuf::from(&target_folder.id);
-        if target_path.starts_with(&old_path) {
-            return Err(String::from(
-                "Folder move failed: cannot move a folder into itself",
-            ));
-        }
-        let Some(folder_name) = old_path.file_name() else {
-            return Err(String::from(
-                "Folder move failed: source folder has no name",
-            ));
-        };
-        let new_path = target_path.join(folder_name);
-        if old_path == new_path {
-            return Ok(FolderDropResult {
-                moved_paths: Vec::new(),
-                status: Some(String::from("Folder move unchanged")),
-            });
-        }
-        if new_path.exists() {
-            return Err(format!(
-                "Folder move failed: {} already exists",
-                folder_name.to_string_lossy()
-            ));
-        }
-        fs::rename(&old_path, &new_path).map_err(|error| format!("Folder move failed: {error}"))?;
-        if let Err(error) = self.relocate_moved_folder(&old_path, &new_path, &target_path) {
-            let _ = fs::rename(&new_path, &old_path);
-            return Err(error);
-        }
-        Ok(FolderDropResult {
-            moved_paths: vec![(old_path, new_path.clone())],
-            status: Some(format!(
-                "Moved folder {}",
-                new_path
-                    .file_name()
-                    .map(|name| name.to_string_lossy().to_string())
-                    .unwrap_or_else(|| new_path.display().to_string())
-            )),
-        })
-    }
-
-    fn move_files_to_folder(
-        &mut self,
-        file_ids: &[String],
-        target_folder_id: &str,
-    ) -> Result<FolderDropResult, String> {
-        if self.rename_active() {
-            return Err(String::from("Finish rename before moving files"));
-        }
-        let target_folder = self
-            .find_folder(target_folder_id)
-            .cloned()
-            .ok_or_else(|| String::from("File move failed: target folder is missing"))?;
-        let target_path = PathBuf::from(&target_folder.id);
-        if !target_path.is_dir() {
-            return Err(String::from("File move failed: target folder is missing"));
-        }
-        let mut moves = Vec::new();
-        let mut seen = HashSet::new();
-        for id in file_ids {
-            if !seen.insert(id.clone()) {
-                continue;
-            }
-            let old_path = PathBuf::from(id);
-            if !old_path.is_file() {
-                return Err(format!(
-                    "File move failed: {} is missing",
-                    old_path
-                        .file_name()
-                        .map(|name| name.to_string_lossy().to_string())
-                        .unwrap_or_else(|| old_path.display().to_string())
-                ));
-            }
-            if old_path.parent() == Some(target_path.as_path()) {
-                continue;
-            }
-            let Some(file_name) = old_path.file_name() else {
-                return Err(String::from("File move failed: source file has no name"));
-            };
-            let new_path = target_path.join(file_name);
-            if new_path.exists() {
-                return Err(format!(
-                    "File move failed: {} already exists",
-                    file_name.to_string_lossy()
-                ));
-            }
-            moves.push((old_path, new_path));
-        }
-        if moves.is_empty() {
-            return Ok(FolderDropResult {
-                moved_paths: Vec::new(),
-                status: Some(String::from("File move unchanged")),
-            });
-        }
-        let mut completed = Vec::new();
-        for (old_path, new_path) in &moves {
-            if let Err(error) = fs::rename(old_path, new_path) {
-                for (moved_old, moved_new) in completed.iter().rev() {
-                    let _ = fs::rename(moved_new, moved_old);
-                }
-                return Err(format!("File move failed: {error}"));
-            }
-            completed.push((old_path.clone(), new_path.clone()));
-        }
-        if let Err(error) = self.relocate_moved_files(&completed, &target_path) {
-            for (moved_old, moved_new) in completed.iter().rev() {
-                let _ = fs::rename(moved_new, moved_old);
-            }
-            return Err(error);
-        }
-        Ok(FolderDropResult {
-            moved_paths: completed.clone(),
-            status: Some(format!(
-                "Moved {} file{}",
-                completed.len(),
-                plural(completed.len())
-            )),
-        })
     }
 }
